+++ src/Builder/FormBuilder.php
@@ -422,6 +422,14 @@ class FormBuilder
             'klarnaStatuses' => [Config::MOLLIE_STATUS_KLARNA_AUTHORIZED, Config::MOLLIE_STATUS_KLARNA_SHIPPED],
             'applePayDirect' => (int) Configuration::get(Config::MOLLIE_APPLE_PAY_DIRECT),
             'applePayDIrectStyle' => (int) Configuration::get(Config::MOLLIE_APPLE_PAY_DIRECT_STYLE),
+            'isBancontactQrCodeEnabled' => (int) Configuration::get(Config::MOLLIE_BANCONTACT_QR_CODE_ENABLED),
+            'isLive' => (int) Configuration::get(Config::MOLLIE_ENVIRONMENT),
+            'bancontactQRCodeDescription' => TagsUtility::ppTags(
+                $this->module->l('Only available with the Live API key. [1]Learn more[/1] about QR Codes.', self::FILE_NAME),
+                [
+                    $this->module->display($this->module->getPathUri(), 'views/templates/admin/mollie_bancontact_qr_code_info.tpl'),
+                ]
+            ),
         ];
 
         return $input;