+++ upgrade/Upgrade-5.2.0.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Mollie       https://www.mollie.nl
+ *
+ * @author      Mollie B.V. <info@mollie.nl>
+ * @copyright   Mollie B.V.
+ * @license     https://github.com/mollie/PrestaShop/blob/master/LICENSE.md
+ *
+ * @see        https://github.com/mollie/PrestaShop
+ */
+
+use Mollie\Config\Config;
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+/**
+ * @param Mollie $module
+ *
+ * @return bool
+ */
+function upgrade_module_5_2_0($module)
+{
+    Configuration::updateValue(Config::MOLLIE_BANCONTACT_QR_CODE_ENABLED, 0);
+
+    return true;
+}